+++ src/views/pages/Dashboard/style.css
@@ -1,4 +1,4 @@
-.landing-container {
+.landing-containe {
     display: flex;
     flex-direction: column;
     align-items: flex-start;
